refactor: Move all ViewModels into the ViewModels folder and namespace
