+++ WpfGui/ViewModels/ImportViewModel.cs
@@ -8,7 +8,7 @@ using AssimilationSoftware.TodoSort.WpfGui.Interfaces;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 
-namespace AssimilationSoftware.TodoSort.WpfGui
+namespace AssimilationSoftware.TodoSort.WpfGui.ViewModels
 {
     public class ImportViewModel : ObservableObject
     {
