+++ WpfGui/ViewModels/RankViewModel.cs
@@ -5,7 +5,7 @@ using AssimilationSoftware.TodoSort.WpfGui.Views;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 
-namespace AssimilationSoftware.TodoSort.WpfGui
+namespace AssimilationSoftware.TodoSort.WpfGui.ViewModels
 {
     public class RankViewModel : ObservableObject
     {
